+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/file/object_storage/ObjectStorageFormattingWriter.kt
@@ -8,7 +8,6 @@ import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.command.object_storage.AvroFormatConfiguration
 import io.airbyte.cdk.load.command.object_storage.CSVFormatConfiguration
 import io.airbyte.cdk.load.command.object_storage.JsonFormatConfiguration
-import io.airbyte.cdk.load.command.object_storage.MSSQLCSVFormatConfiguration
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfigurationProvider
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageFormatConfigurationProvider
 import io.airbyte.cdk.load.command.object_storage.ParquetFormatConfiguration
@@ -16,7 +15,6 @@ import io.airbyte.cdk.load.data.ObjectType
 import io.airbyte.cdk.load.data.avro.AvroMapperPipelineFactory
 import io.airbyte.cdk.load.data.avro.toAvroRecord
 import io.airbyte.cdk.load.data.avro.toAvroSchema
-import io.airbyte.cdk.load.data.csv.MSSQLCsvRowValidator
 import io.airbyte.cdk.load.data.csv.toCsvRecord
 import io.airbyte.cdk.load.data.dataWithAirbyteMeta
 import io.airbyte.cdk.load.data.parquet.ParquetMapperPipelineTest
@@ -43,12 +41,16 @@ interface ObjectStorageFormattingWriter : Closeable {
     fun flush()
 }
 
+interface ObjectStorageFormattingWriterFactory {
+    fun create(stream: DestinationStream, outputStream: OutputStream): ObjectStorageFormattingWriter
+}
+
 @Singleton
 @Secondary
-class ObjectStorageFormattingWriterFactory(
+class DefaultObjectStorageFormattingWriterFactory(
     private val formatConfigProvider: ObjectStorageFormatConfigurationProvider,
-) {
-    fun create(
+) : ObjectStorageFormattingWriterFactory {
+    override fun create(
         stream: DestinationStream,
         outputStream: OutputStream
     ): ObjectStorageFormattingWriter {
@@ -74,14 +76,6 @@ class ObjectStorageFormattingWriterFactory(
                     flatten
                 )
             is CSVFormatConfiguration -> CSVFormattingWriter(stream, outputStream, flatten)
-            is MSSQLCSVFormatConfiguration ->
-                MSSQLCSVFormattingWriter(
-                    stream,
-                    outputStream,
-                    (formatConfigProvider.objectStorageFormatConfiguration
-                            as MSSQLCSVFormatConfiguration)
-                        .validateValuesPreLoad,
-                )
         }
     }
 }
@@ -105,31 +99,6 @@ class JsonFormattingWriter(
 
     override fun close() {
         outputStream.close()
-    }
-}
-
-class MSSQLCSVFormattingWriter(
-    private val stream: DestinationStream,
-    outputStream: OutputStream,
-    validateValuesPreLoad: Boolean,
-) : ObjectStorageFormattingWriter {
-    private val finalSchema = stream.schema.withAirbyteMeta(true)
-    private val printer = finalSchema.toCsvPrinterWithHeader(outputStream)
-    private val mssqlRowValidator = MSSQLCsvRowValidator(validateValuesPreLoad)
-    override fun accept(record: DestinationRecordAirbyteValue) {
-        printer.printRecord(
-            mssqlRowValidator
-                .validate(record, this.finalSchema)
-                .dataWithAirbyteMeta(stream, true)
-                .toCsvRecord(finalSchema),
-        )
-    }
-    override fun flush() {
-        printer.flush()
-    }
-
-    override fun close() {
-        printer.close()
     }
 }
