+++ airbyte-integrations/connectors/destination-mssql/src/main/kotlin/io/airbyte/integrations/destination/mssql/v2/MSSQLBulkLoadStreamLoader.kt
@@ -7,11 +7,8 @@ package io.airbyte.integrations.destination.mssql.v2
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings
 import io.airbyte.cdk.load.command.Dedupe
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.command.object_storage.MSSQLCSVFormatConfiguration
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfiguration
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfigurationProvider
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageFormatConfiguration
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageFormatConfigurationProvider
 import io.airbyte.cdk.load.command.object_storage.ObjectStoragePathConfiguration
 import io.airbyte.cdk.load.command.object_storage.ObjectStoragePathConfigurationProvider
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageUploadConfiguration
@@ -21,7 +18,6 @@ import io.airbyte.cdk.load.file.DefaultTimeProvider
 import io.airbyte.cdk.load.file.NoopProcessor
 import io.airbyte.cdk.load.file.azureBlobStorage.AzureBlobClient
 import io.airbyte.cdk.load.file.object_storage.BufferedFormattingWriterFactory
-import io.airbyte.cdk.load.file.object_storage.ObjectStorageFormattingWriterFactory
 import io.airbyte.cdk.load.file.object_storage.ObjectStoragePathFactory
 import io.airbyte.cdk.load.message.Batch
 import io.airbyte.cdk.load.message.object_storage.LoadedObject
@@ -86,7 +82,7 @@ class MSSQLBulkLoadStreamLoader(
     override suspend fun createBatchAccumulator(): BatchAccumulator {
         val writerFactory =
             BufferedFormattingWriterFactory(
-                ObjectStorageFormattingWriterFactory(CsvFormatProvider(validateValuesPreLoad)),
+                MssqlObjectStorageFormattingWriterFactory(validateValuesPreLoad),
                 NoOpCompressionProvider()
             )
 
@@ -185,13 +181,6 @@ class MSSQLBulkLoadStreamLoader(
         ObjectStorageCompressionConfigurationProvider<ByteArrayOutputStream> {
         override val objectStorageCompressionConfiguration =
             ObjectStorageCompressionConfiguration(compressor = NoopProcessor)
-    }
-
-    /** Provides a CSV format configuration for object storage. */
-    private class CsvFormatProvider(validateValuesPreLoad: Boolean) :
-        ObjectStorageFormatConfigurationProvider {
-        override val objectStorageFormatConfiguration: ObjectStorageFormatConfiguration =
-            MSSQLCSVFormatConfiguration(validateValuesPreLoad = validateValuesPreLoad)
     }
 
     /** Defines how blob paths are constructed. For example, "blob/.../{timestamp}/..." etc. */
